numpad file refactoring
